Test 00 (#7)
* fix

* update assets

* experiment

* test_00

* ready

* oops

* ip
